add the 'id' as a column to check in the Grammar when parameterizing
The Neo4j client have the ability to replace properties b/w { } with
passed values. We’ve already implemented a way to replace the id(n)
that is treated differently by Neo4j to get a node by its ‘id’, but
when the dev decides to fetch the ‘id’ within specified columns i.e.
User::first([‘id’, ‘name’]), this also has to be treated differently
through the Client so this commit does it :smile: