add support for 'many' mutations to the Eloquent Builder
This will allow us to have two types of mutations, either coming from a
HasOne or belongsTo relationship or form a hasMany relationship where
dealing with related results is much different than single model
results, with many results we are dealing with an Eloquent\Collection
instead of a plain array.

Also adds a lot of mutation-specific helper functions that makes it
easy to work with mutations in a centralized manner.