make the abstract Edges\Relation inherit Delegate and use Finder
With the introduction of the Delegate and Finder classes they have
taken most of the hassle of dealing with the database client and most
common tasks away from Relation. Also introduce new functionality in
Relation that stores and returns Parent and Related models of the
relationship.

This commit closes an issue where setting a relation on the inheriting
class does not properly set all the aspects of the relation including
Start and End nodes, Related model and its attributes and database
connection.