add whereCarried() and with() to Query Builder
whereCarried() and with() methods are meant to work with the WITH
clause of a query where whereCarried() will add the identifier carried
from the with() function that fills up the $with property of the query.
Once the compiler hit a WHERE clause of the type Carried it will not
wrap it or change anything about the column key.