rid illuminate/database from being a dependency
NeoEloquent will no longer depend on illuminate/database, the code has
been copied into NE and is now a proprietary part of it. This way we
won’t have to maintain multiple branches so that we keep compatibility
with Laravel, and NE can be used with any version of Laravel.